requirements.txt
streamlit
langchain
openai
newsapi-python
python-dotenv
pytest
tenacity